04_existing_resources/requirements.txt
aws-cdk-lib==2.142.1
constructs>=10.0.0,<11.0.0
python-dotenv